Safari-609.1.11.1/LayoutTests/security/contentSecurityPolicy/plugins-types-allows-quicktime-plugin-replacement-expected.txt

ALERT: PASS did load plugin.
